+++ rpl_wei/wei_workflow_base.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, List, Optional, Tuple
 from devtools import debug
 
 from rpl_wei.data_classes import Module, WorkCell, Workflow
-from rpl_wei.executors import StepExecutor
+from rpl_wei.executors.step_executor import StepExecutor
 from rpl_wei.loggers import WEI_Logger
 from rpl_wei.validators import ModuleValidator, StepValidator
